Use stream instead of buffers to avoid big memory allocations
Previously we were getting the error 'RangeError: Array buffer
allocation failed', because we were concatenating the stream info
to an array, and when the stream was finished, converting it into a
Buffer.

This forced us to allocate twice (one for the array and another for the
Buffer), and to run out of memory in the container.

At the beginning I tried implementing it using streams, but it didn't
work, apparently the new S3 library can handle it now.